Move code between test fixtures and add global start/finish functions to `conftest.py`.
The session scoped fixture in fixtures.py, decorated with:
```
@pytest.fixture(scope="session")
```
is not executed so move code to conftest.pytest_sessionstart().

It would be faster to delete table data after each test rather than destroying
and recreating the tables as done now, but the code to delete table data
is not working.  The `session` fixture still destroys and recreates tables.